+++ native/omnisus-db-dbf/fuzz/fuzz_targets/dbc.rs
@@ -1,0 +1,11 @@
+#![no_main]
+
+use _native::dbc::decompress;
+use libfuzzer_sys::fuzz_target;
+
+fuzz_target!(|data: &[u8]| {
+    if data.len() > 64 * 1024 {
+        return;
+    }
+    let _ = decompress(data);
+});